Proyecto Farmacia Zarcos/zarcos_be/requirements.txt
# pip freeze
Django 
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
numpy

# pip install -r requirenebts.txt